Added non-exploding flak flakspec commands
noexplosion=x - Setting this to 1 will turn off explosion for the flak object and stop the explosion radius ( will act like a bullet from normal weapon )

stick=x - Setting this to 1 will allow for an object to stick out of a static level object when it collides with it (arrow hits a wall, arrow sticks out of wall). stick.x and stick_d2.tga (this will before the model that sticks out of the static object it hit) must be in the appropriate flak folder.

sticklife=x - The amount of time in milliseconds that the stick object will stay "alive for". After this amount of time is up the object will disappear.Added non-exploding flak
flakspec commands

noexplosion=x - Setting this to 1 will turn off explosion for the flak object and stop the explosion radius ( will act like a bullet from normal weapon )

stick=x - Setting this to 1 will allow for an object to stick out of a static level object when it collides with it (arrow hits a wall, arrow sticks out of wall). stick.x and stick_d2.tga (this will before the model that sticks out of the static object it hit) must be in the appropriate flak folder.

sticklife=x - The amount of time in milliseconds that the stick object will stay "alive for". After this amount of time is up the object will disappear.

Fixed bounceonhit=x - This command was intended to allow for X to be the amount of bounces, but never actually worked that way until now